refactor [types/schemas] put all types and schemas in their own folders
