+++ src/modules/enrollment/enrollment.router.ts
@@ -1,6 +1,5 @@
 import { Router } from "express";
 import { authenticateToken } from "../../middleware/authenticationMiddleware";
-
 import { completeProfile } from "./enrollment.controller";
 
 const enrollmentRouter = Router();
