+++ src/modules/enrollment/enrollment.service.ts
@@ -1,7 +1,8 @@
 import { enrollmentRepository } from "../../repositories";
-import { newEnrollmentBody } from "./enrollment.types";
+import { newEnrollmentBody } from "../../types/enrollment.types";
 
 async function createEnrollment(body: newEnrollmentBody) {    
+    //TODO: include verification "if there is already an enrollment"
     await enrollmentRepository.create(body);
     return;
 }
