+++ src/modules/users/user.router.ts
@@ -1,5 +1,4 @@
 import { Router } from "express";
-
 import { signUpUser } from "./users.controller";
 
 const userRouter = Router();
