+++ src/modules/users/users.service.ts
@@ -1,22 +1,9 @@
 import bcrypt from "bcrypt";
-import joi from "joi";
 import { duplicatedEmailError } from "../../errors/duplicatedEmailError";
 import { invalidPasswordFormatError } from "../../errors/invalidPasswordFormatError";
 import { userRepository } from "../../repositories";
-
-type newUserBody = {
-    email: string,
-    password: string
-}
-
-const passwordSchema = joi.object({
-    password: joi
-        .string()
-        .trim()
-        .required()
-        .min(8)
-        .regex(new RegExp("^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\\$%\\^&\\*])(?=.{8,})")) //at least 1 lower, 1 upper letter and a special char
-});
+import { passwordSchema } from "../../schemas/user.schemas";
+import { newUserBody } from "../../types/users.type";
 
 async function userSignUp({ email, password }: newUserBody) {
     await validatePasswordFormatOrFail(password);
