+++ src/types/users.type.ts
@@ -1,0 +1,6 @@
+export type newUserBody = {
+  email: string,
+  password: string
+};
+
+export type userBody = newUserBody;
